Fixed some compiler warnings on MSVC 64bit
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@1529 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/add_on/scriptbuilder/scriptbuilder.cpp
@@ -174,7 +174,7 @@ int CScriptBuilder::LoadScriptSection(const char *filename)
 		return -1;
 	}
 
-	return ProcessScriptSection(code.c_str(), code.length(), filename);
+	return ProcessScriptSection(code.c_str(), (unsigned int)(code.length()), filename);
 }
 
 int CScriptBuilder::ProcessScriptSection(const char *script, unsigned int length, const char *sectionname)
